+++ include/system/operatingsystem.h
@@ -1,0 +1,18 @@
+#ifndef OPERATINGSYSTEM_H
+#define OPERATINGSYSTEM_H
+
+namespace Nickvision::System
+{
+    /**
+     * @brief Supported operating system values.
+     */
+    enum class OperatingSystem
+    {
+        Windows,
+        Linux,
+        MacOS,
+        Other
+    };
+}
+
+#endif //OPERATINGSYSTEM_H